# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface PlaceIdRequest {
  /** @minLength 1 */
  placeId: string;
}

export interface SavedPlacesEnvelope {
  placeIds: string[];
}

export interface CreateReviewRequest {
  /**
     * @minimum 1
     * @maximum 5
     */
  overall: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  communication: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  staffUnderstanding: number;
  /**
     * @minLength 5
     * @maxLength 1000
     */
  text: string;
}

export interface Review {
  id: string;
  userName: string;
  overall: number;
  communication: number;
  staffUnderstanding: number;
  text: string;
  createdAt: string;
}

export interface ReviewsEnvelope {
  reviews: Review[];
}

export interface HealthStatus {
  status: string;
}