# lib/api-zod/src/generated/types/createReviewRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateReviewRequest {
  /**
     * @minimum 1
     * @maximum 5
     */
  overall: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  communication: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  staffUnderstanding: number;
  /**
     * @minLength 5
     * @maxLength 1000
     */
  text: string;
}