# lib/api-zod/src/generated/types/reviewsEnvelope.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Review } from './review';

export interface ReviewsEnvelope {
  reviews: Review[];
}